+++ spec/Pim/Bundle/FilterBundle/Filter/CategoryFilterSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\FilterBundle\Filter\Product;
+namespace spec\Pim\Bundle\FilterBundle\Filter;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Manager\ProductCategoryManager;
@@ -12,9 +12,12 @@ use Symfony\Component\Form\FormFactoryInterface;
 
 class CategoryFilterSpec extends ObjectBehavior
 {
-    function let(FormFactoryInterface $factory, ProductFilterUtility $utility, ProductCategoryManager $manager)
-    {
-        $this->beConstructedWith($factory, $utility, $manager);
+    function let(
+        FormFactoryInterface $factory,
+        ProductFilterUtility $utility,
+        CategoryRepositoryInterface $categoryRepo
+    ) {
+        $this->beConstructedWith($factory, $utility, $categoryRepo);
     }
 
     function it_is_an_oro_number_filter()
@@ -22,38 +25,32 @@ class CategoryFilterSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf('Oro\Bundle\FilterBundle\Filter\NumberFilter');
     }
 
-    function it_applies_a_filter_on_all_products(
-        FilterDatasourceAdapterInterface $datasource,
-        $utility
-    ) {
+    function it_applies_a_filter_on_all_products(FilterDatasourceAdapterInterface $datasource)
+    {
         $this->apply($datasource, ['value' => ['categoryId' => -2]])->shouldReturn(true);
     }
 
     function it_applies_a_filter_by_unclassified_products(
-        FilterDatasourceAdapterInterface $datasource,
         $utility,
-        $manager,
-        CategoryRepositoryInterface $repo,
+        $categoryRepo,
+        FilterDatasourceAdapterInterface $datasource,
         CategoryInterface $tree
     ) {
-        $manager->getCategoryRepository()->willReturn($repo);
         $tree->getId()->willReturn(1);
-        $repo->find(1)->willReturn($tree);
-        $repo->getAllChildrenIds($tree)->willReturn([2, 3]);
+        $categoryRepo->find(1)->willReturn($tree);
+        $categoryRepo->getAllChildrenIds($tree)->willReturn([2, 3]);
         $utility->applyFilter($datasource, 'categories.id', 'NOT IN', [2, 3])->shouldBeCalled();
 
         $this->apply($datasource, ['value' => ['categoryId' => -1, 'treeId' => 1]]);
     }
 
     function it_applies_a_filter_by_in_category(
-        FilterDatasourceAdapterInterface $datasource,
         $utility,
-        $manager,
-        CategoryRepositoryInterface $repo,
+        $categoryRepo,
+        FilterDatasourceAdapterInterface $datasource,
         CategoryInterface $category
     ) {
-        $manager->getCategoryRepository()->willReturn($repo);
-        $repo->find(42)->willReturn($category);
+        $categoryRepo->find(42)->willReturn($category);
         $category->getId()->willReturn(42);
         $utility->applyFilter($datasource, 'categories.id', 'IN', [42])->shouldBeCalled();
 
@@ -61,17 +58,15 @@ class CategoryFilterSpec extends ObjectBehavior
     }
 
     function it_applies_a_filter_by_in_category_with_children(
-        FilterDatasourceAdapterInterface $datasource,
         $utility,
-        $manager,
-        CategoryRepositoryInterface $repo,
+        $categoryRepo,
+        FilterDatasourceAdapterInterface $datasource,
         CategoryInterface $category
     ) {
-        $manager->getCategoryRepository()->willReturn($repo);
-        $repo->find(42)->willReturn($category);
+        $categoryRepo->find(42)->willReturn($category);
         $category->getId()->willReturn(42);
-        $repo->find(42)->willReturn($category);
-        $repo->getAllChildrenIds($category)->willReturn([2, 3]);
+        $categoryRepo->find(42)->willReturn($category);
+        $categoryRepo->getAllChildrenIds($category)->willReturn([2, 3]);
 
         $utility->applyFilter($datasource, 'categories.id', 'IN', [2, 3, 42])->shouldBeCalled();