+++ spec/Pim/Component/Classification/Updater/CategoryUpdaterSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Component\Catalog\Updater;
 use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\CategoryTranslation;
-use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
+use Pim\Component\Classification\Model\CategoryInterface;
 use Prophecy\Argument;
 
 class CategoryUpdaterSpec extends ObjectBehavior
@@ -17,7 +17,7 @@ class CategoryUpdaterSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Pim\Component\Catalog\Updater\CategoryUpdater');
+        $this->shouldHaveType('Pim\Component\Classification\Updater\CategoryUpdater');
     }
 
     function it_is_a_updater()