+++ src/Pim/Bundle/ClassificationBundle/DependencyInjection/PimClassificationExtension.php
@@ -2,7 +2,9 @@
 
 namespace Pim\Bundle\ClassificationBundle\DependencyInjection;
 
+use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 /**
@@ -19,5 +21,8 @@ class PimClassificationExtension extends Extension
      */
     public function load(array $configs, ContainerBuilder $container)
     {
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__ .'/../Resources/config'));
+        $loader->load('factories.yml');
+        $loader->load('updaters.yml');
     }
 }