+++ src/Pim/Bundle/EnrichBundle/Controller/CategoryTreeController.php
@@ -4,6 +4,7 @@ namespace Pim\Bundle\EnrichBundle\Controller;
 
 use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
 use Akeneo\Component\StorageUtils\Saver\SaverInterface;
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Oro\Bundle\SecurityBundle\Annotation\AclAncestor;
 use Oro\Bundle\SecurityBundle\SecurityFacade;
@@ -12,12 +13,15 @@ use Pim\Bundle\CatalogBundle\Model\CategoryInterface;
 use Pim\Bundle\EnrichBundle\AbstractController\AbstractDoctrineController;
 use Pim\Bundle\EnrichBundle\Event\CategoryEvents;
 use Pim\Bundle\UserBundle\Context\UserContext;
+use Pim\Component\Classification\Factory\CategoryFactory;
+use Pim\Component\Classification\Repository\CategoryRepositoryInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -50,23 +54,31 @@ class CategoryTreeController extends AbstractDoctrineController
     /** @var RemoverInterface */
     protected $categoryRemover;
 
+    /** @var CategoryFactory */
+    protected $categoryFactory;
+
+    /** @var CategoryRepositoryInterface */
+    protected $categoryRepository;
+
     /**
      * Constructor
      *
-     * @param Request                  $request
-     * @param EngineInterface          $templating
-     * @param RouterInterface          $router
-     * @param SecurityContextInterface $securityContext
-     * @param FormFactoryInterface     $formFactory
-     * @param ValidatorInterface       $validator
-     * @param TranslatorInterface      $translator
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param ManagerRegistry          $doctrine
-     * @param CategoryManager          $categoryManager
-     * @param UserContext              $userContext
-     * @param SecurityFacade           $securityFacade
-     * @param SaverInterface           $categorySaver
-     * @param RemoverInterface         $categoryRemover
+     * @param Request                     $request
+     * @param EngineInterface             $templating
+     * @param RouterInterface             $router
+     * @param SecurityContextInterface    $securityContext
+     * @param FormFactoryInterface        $formFactory
+     * @param ValidatorInterface          $validator
+     * @param TranslatorInterface         $translator
+     * @param EventDispatcherInterface    $eventDispatcher
+     * @param ManagerRegistry             $doctrine
+     * @param CategoryManager             $categoryManager
+     * @param UserContext                 $userContext
+     * @param SecurityFacade              $securityFacade
+     * @param SaverInterface              $categorySaver
+     * @param RemoverInterface            $categoryRemover
+     * @param CategoryFactory             $categoryFactory
+     * @param CategoryRepositoryInterface $categoryRepository
      */
     public function __construct(
         Request $request,
@@ -82,7 +94,9 @@ class CategoryTreeController extends AbstractDoctrineController
         UserContext $userContext,
         SecurityFacade $securityFacade,
         SaverInterface $categorySaver,
-        RemoverInterface $categoryRemover
+        RemoverInterface $categoryRemover,
+        CategoryFactory $categoryFactory,
+        CategoryRepositoryInterface $categoryRepository
     ) {
         parent::__construct(
             $request,
@@ -96,11 +110,13 @@ class CategoryTreeController extends AbstractDoctrineController
             $doctrine
         );
 
-        $this->categoryManager = $categoryManager;
-        $this->userContext     = $userContext;
-        $this->securityFacade  = $securityFacade;
-        $this->categorySaver   = $categorySaver;
-        $this->categoryRemover = $categoryRemover;
+        $this->categoryManager    = $categoryManager;
+        $this->userContext        = $userContext;
+        $this->securityFacade     = $securityFacade;
+        $this->categorySaver      = $categorySaver;
+        $this->categoryRemover    = $categoryRemover;
+        $this->categoryFactory    = $categoryFactory;
+        $this->categoryRepository = $categoryRepository;
     }
 
     /**
@@ -118,18 +134,20 @@ class CategoryTreeController extends AbstractDoctrineController
     public function listTreeAction(Request $request)
     {
         $selectNodeId = $request->get('select_node_id', -1);
+        $relatedEntity = $this->getRequest()->get('related_entity', 'product');
+
         try {
             $selectNode = $this->findCategory($selectNodeId);
         } catch (NotFoundHttpException $e) {
-            $selectNode = $this->userContext->getUserTree();
+            $selectNode = $this->userContext->getUserCategoryTree($relatedEntity);
         }
 
         return [
-            'trees'          => $this->categoryManager->getTrees(),
+            'trees'          => $this->categoryRepository->getTrees(),
             'selectedTreeId' => $selectNode->isRoot() ? $selectNode->getId() : $selectNode->getRoot(),
             'include_sub'    => (bool) $this->getRequest()->get('include_sub', false),
-            'product_count'  => (bool) $this->getRequest()->get('with_products_count', true),
-            'related_entity' => $this->getRequest()->get('related_entity', 'product')
+            'item_count'  => (bool) $this->getRequest()->get('with_items_count', true),
+            'related_entity' => $relatedEntity
         ];
     }
 
@@ -148,6 +166,7 @@ class CategoryTreeController extends AbstractDoctrineController
         $parentId      = $request->get('parent');
         $prevSiblingId = $request->get('prev_sibling');
 
+        // TODO: Change this in PIM-4409
         $this->categoryManager->move($categoryId, $parentId, $prevSiblingId);
 
         return new JsonResponse(['status' => 1]);
@@ -176,28 +195,28 @@ class CategoryTreeController extends AbstractDoctrineController
             return ['categories' => []];
         }
 
-        $selectNodeId      = $this->getRequest()->get('select_node_id', -1);
-        $withProductsCount = (bool) $this->getRequest()->get('with_products_count', false);
-        $includeParent     = (bool) $this->getRequest()->get('include_parent', false);
-        $includeSub        = (bool) $this->getRequest()->get('include_sub', false);
-        $relatedEntity     = $this->getRequest()->get('related_entity', 'product');
+        $selectNodeId   = $this->getRequest()->get('select_node_id', -1);
+        $withItemsCount = (bool)$this->getRequest()->get('with_items_count', false);
+        $includeParent  = (bool)$this->getRequest()->get('include_parent', false);
+        $includeSub     = (bool)$this->getRequest()->get('include_sub', false);
+        $relatedEntity  = $this->getRequest()->get('related_entity', 'product');
 
         try {
             $selectNode = $this->findCategory($selectNodeId);
 
-            if (!$this->categoryManager->isAncestor($parent, $selectNode)) {
+            if (!$this->categoryRepository->isAncestor($parent, $selectNode)) {
                 $selectNode = null;
             }
         } catch (NotFoundHttpException $e) {
             $selectNode = null;
         }
 
-        if ($selectNode !== null) {
-            $categories = $this->getChildren($parent->getId(), $selectNode->getId());
-            $view = 'PimEnrichBundle:CategoryTree:children-tree.json.twig';
-        } else {
-            $categories = $this->getChildren($parent->getId());
+        $categories = $this->getChildrenCategories($request, $selectNode);
+
+        if (null === $selectNode) {
             $view = 'PimEnrichBundle:CategoryTree:children.json.twig';
+        } else {
+            $view = 'PimEnrichBundle:CategoryTree:children-tree.json.twig';
         }
 
         return $this->render(
@@ -206,7 +225,7 @@ class CategoryTreeController extends AbstractDoctrineController
                 'categories'     => $categories,
                 'parent'         => ($includeParent) ? $parent : null,
                 'include_sub'    => $includeSub,
-                'product_count'  => $withProductsCount,
+                'item_count'     => $withItemsCount,
                 'select_node'    => $selectNode,
                 'related_entity' => $relatedEntity
             ],
@@ -239,10 +258,10 @@ class CategoryTreeController extends AbstractDoctrineController
     {
         if ($parent) {
             $parent = $this->findCategory($parent);
-            $category = $this->categoryManager->getCategoryInstance();
+            $category = $this->categoryFactory->create();
             $category->setParent($parent);
         } else {
-            $category = $this->categoryManager->getTreeInstance();
+            $category = $this->categoryFactory->create();
         }
 
         $category->setCode($request->get('label'));
@@ -336,7 +355,7 @@ class CategoryTreeController extends AbstractDoctrineController
      */
     protected function findCategory($categoryId)
     {
-        $category = $this->categoryManager->getEntityRepository()->find($categoryId);
+        $category = $this->categoryRepository->find($categoryId);
 
         if (!$category) {
             throw $this->createNotFoundException('Category not found');
@@ -358,13 +377,21 @@ class CategoryTreeController extends AbstractDoctrineController
     }
 
     /**
-     * @param int   $parentId
-     * @param mixed $selectNodeId
+     * @param Request                $request
+     * @param CategoryInterface|null $selectNode
      *
-     * @return CategoryInterface[]
+     * @return array|ArrayCollection
      */
-    protected function getChildren($parentId, $selectNodeId = false)
+    protected function getChildrenCategories(Request $request, $selectNode)
     {
-        return $this->categoryManager->getChildren($parentId, $selectNodeId);
+        $parent = $this->findCategory($request->get('id'));
+
+        if (null !== $selectNode) {
+            $categories = $this->categoryRepository->getChildrenTreeByParentId($parent->getId(), $selectNode->getId());
+        } else {
+            $categories = $this->categoryRepository->getChildrenByParentId($parent->getId());
+        }
+
+        return $categories;
     }
 }