+++ src/Pim/Bundle/EnrichBundle/Controller/ProductController.php
@@ -22,6 +22,7 @@ use Pim\Bundle\EnrichBundle\Event\ProductEvents;
 use Pim\Bundle\EnrichBundle\Manager\SequentialEditManager;
 use Pim\Bundle\UserBundle\Context\UserContext;
 use Pim\Bundle\VersioningBundle\Manager\VersionManager;
+use Pim\Component\Classification\Factory\CategoryFactory;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\EventDispatcher\Event;
@@ -80,6 +81,9 @@ class ProductController extends AbstractDoctrineController
     /** @var ProductBuilderInterface */
     protected $productBuilder;
 
+    /** @var CategoryFactory */
+    protected $categoryFactory;
+
     /**
      * Constant used to redirect to the datagrid when save edit form
      *
@@ -131,6 +135,7 @@ class ProductController extends AbstractDoctrineController
      * @param SequentialEditManager    $seqEditManager
      * @param RemoverInterface         $productRemover
      * @param ProductBuilderInterface  $productBuilder
+     * @param CategoryFactory          $categoryFactory
      */
     public function __construct(
         Request $request,
@@ -152,7 +157,8 @@ class ProductController extends AbstractDoctrineController
         MediaManager $mediaManager,
         SequentialEditManager $seqEditManager,
         RemoverInterface $productRemover,
-        ProductBuilderInterface $productBuilder
+        ProductBuilderInterface $productBuilder,
+        CategoryFactory $categoryFactory
     ) {
         parent::__construct(
             $request,
@@ -177,6 +183,7 @@ class ProductController extends AbstractDoctrineController
         $this->seqEditManager    = $seqEditManager;
         $this->productRemover    = $productRemover;
         $this->productBuilder    = $productBuilder;
+        $this->categoryFactory   = $categoryFactory;
     }
 
     /**
@@ -437,7 +444,7 @@ class ProductController extends AbstractDoctrineController
     public function listCategoriesAction(Request $request, $id, $categoryId)
     {
         $product = $this->findProductOr404($id);
-        $parent = $this->findOr404($this->categoryManager->getCategoryClass(), $categoryId);
+        $parent = $this->findOr404($this->categoryFactory->getCategoryClass(), $categoryId);
         $categories = null;
 
         $includeParent = $request->get('include_parent', false);