+++ src/Pim/Component/Classification/Repository/CategoryRepositoryInterface.php
@@ -140,4 +140,19 @@ interface CategoryRepositoryInterface extends
      * @return array - list of Nodes in path
      */
     public function getPath($node);
+
+    /**
+     * @return array
+     */
+    public function getTrees();
+
+    /**
+     * Check if a parent node is an ancestor of a child node
+     *
+     * @param CategoryInterface $parentNode
+     * @param CategoryInterface $childNode
+     *
+     * @return bool
+     */
+    public function isAncestor(CategoryInterface $parentNode, CategoryInterface $childNode);
 }